+++ internal/release/main.go
@@ -9,6 +9,34 @@ import (
 	"expvar"
 )
 
+const (
+	// NAME is the name of this application
+	NAME = "circonus-logwatch"
+	// ENVPREFIX is the environment variable prefix
+	ENVPREFIX = "CLW"
+)
+
+// vars are manipulated at link time (see goreleaser)
+var (
+	// COMMIT of relase in git repo
+	COMMIT = "none"
+	// DATE of release
+	DATE = "unknown"
+	// TAG of release
+	TAG = ""
+	// VERSION of the release
+	VERSION = "dev"
+)
+
+// Info contains release information
+type Info struct {
+	Name      string
+	Version   string
+	Commit    string
+	BuildDate string
+	Tag       string
+}
+
 func init() {
 	expvar.Publish("app", expvar.Func(info))
 }
